refactor: move native LLVM C++ API code from llvm.h into new llvm.cpp
